Moved lists to pkg/lists
