+++ pkg/lists/list_test.go
@@ -4,7 +4,7 @@
 //
 // SPDX-License-Identifier: BSD-3-Clause
 
-package typ
+package lists
 
 import "testing"
 
@@ -63,7 +63,7 @@ func checkListPointers[T any](t *testing.T, l *List[T], es []*Element[T]) {
 }
 
 func TestList(t *testing.T) {
-	sl := NewList[string]()
+	sl := New[string]()
 	checkListPointers(t, sl, []*Element[string]{})
 
 	// Single element list
@@ -77,7 +77,7 @@ func TestList(t *testing.T) {
 	checkListPointers(t, sl, []*Element[string]{})
 
 	// Bigger list
-	il := NewList[int]()
+	il := New[int]()
 	e2 := il.PushFront(2)
 	e1 := il.PushFront(1)
 	e3 := il.PushBack(3)
@@ -160,8 +160,8 @@ func checkList[T int](t *testing.T, l *List[T], es []T) {
 }
 
 func TestExtending(t *testing.T) {
-	l1 := NewList[int]()
-	l2 := NewList[int]()
+	l1 := New[int]()
+	l2 := New[int]()
 
 	l1.PushBack(1)
 	l1.PushBack(2)
@@ -170,13 +170,13 @@ func TestExtending(t *testing.T) {
 	l2.PushBack(4)
 	l2.PushBack(5)
 
-	l3 := NewList[int]()
+	l3 := New[int]()
 	l3.PushBackList(l1)
 	checkList(t, l3, []int{1, 2, 3})
 	l3.PushBackList(l2)
 	checkList(t, l3, []int{1, 2, 3, 4, 5})
 
-	l3 = NewList[int]()
+	l3 = New[int]()
 	l3.PushFrontList(l2)
 	checkList(t, l3, []int{4, 5})
 	l3.PushFrontList(l1)
@@ -185,19 +185,19 @@ func TestExtending(t *testing.T) {
 	checkList(t, l1, []int{1, 2, 3})
 	checkList(t, l2, []int{4, 5})
 
-	l3 = NewList[int]()
+	l3 = New[int]()
 	l3.PushBackList(l1)
 	checkList(t, l3, []int{1, 2, 3})
 	l3.PushBackList(l3)
 	checkList(t, l3, []int{1, 2, 3, 1, 2, 3})
 
-	l3 = NewList[int]()
+	l3 = New[int]()
 	l3.PushFrontList(l1)
 	checkList(t, l3, []int{1, 2, 3})
 	l3.PushFrontList(l3)
 	checkList(t, l3, []int{1, 2, 3, 1, 2, 3})
 
-	l3 = NewList[int]()
+	l3 = New[int]()
 	l1.PushBackList(l3)
 	checkList(t, l1, []int{1, 2, 3})
 	l1.PushFrontList(l3)
@@ -205,7 +205,7 @@ func TestExtending(t *testing.T) {
 }
 
 func TestRemove(t *testing.T) {
-	l := NewList[int]()
+	l := New[int]()
 	e1 := l.PushBack(1)
 	e2 := l.PushBack(2)
 	checkListPointers(t, l, []*Element[int]{e1, e2})
@@ -217,11 +217,11 @@ func TestRemove(t *testing.T) {
 }
 
 func TestIssue4103(t *testing.T) {
-	l1 := NewList[int]()
+	l1 := New[int]()
 	l1.PushBack(1)
 	l1.PushBack(2)
 
-	l2 := NewList[int]()
+	l2 := New[int]()
 	l2.PushBack(3)
 	l2.PushBack(4)
 
@@ -238,7 +238,7 @@ func TestIssue4103(t *testing.T) {
 }
 
 func TestIssue6349(t *testing.T) {
-	l := NewList[int]()
+	l := New[int]()
 	l.PushBack(1)
 	l.PushBack(2)
 
@@ -256,7 +256,7 @@ func TestIssue6349(t *testing.T) {
 }
 
 func TestMove(t *testing.T) {
-	l := NewList[int]()
+	l := New[int]()
 	e1 := l.PushBack(1)
 	e2 := l.PushBack(2)
 	e3 := l.PushBack(3)
